updates to code
reformat of ut bot to work over series and not over dataframes, and fully writing stc oscilator
please notice that ut bot has been rewritten in a way that allows to get the signals that the indian guy's tutorial mentioned.
at the end of the day that is all we need from it in the matter of getting the data, the only other thing about it is to understand and explain it.